refactor(doctor): extract shell/version/legacy checks + their helpers
