+++ src/core/doctor/checks/mod.rs
@@ -1,9 +1,15 @@
 //! Individual doctor checks, one per file, registered via [`default_checks`].
 
 mod home;
+mod legacy;
+mod shell;
 mod symlink;
 mod uv;
+mod version;
 mod virtualenv;
+
+#[cfg(test)]
+mod test_support;
 
 use super::types::Check;
 
@@ -14,9 +20,8 @@ pub(super) fn default_checks() -> Vec<Box<dyn Check>> {
         Box::new(home::HomeCheck),
         Box::new(virtualenv::VirtualenvCheck),
         Box::new(symlink::SymlinkCheck),
-        // TEMPORARY until Task 5 — still defined in mod.rs:
-        Box::new(super::ShellCheck),
-        Box::new(super::VersionCheck),
-        Box::new(super::LegacyCheck),
+        Box::new(shell::ShellCheck),
+        Box::new(version::VersionCheck),
+        Box::new(legacy::LegacyCheck),
     ]
 }
